Correções na indentação de código-fonte das classes de cidades, e-mails, enderecos, estados, telefones e usuários. Também foi alteradas as classes app.blade.php e home.blade.php e criação da classe flash-message.blade.php.

+++ app/Http/Controllers/UsuariosController.php
@@ -137,6 +137,14 @@ class UsuariosController extends Controller
      */
     public function destroy($id)
     {
+        $usuarios = User::get();
+
+        if ($usuarios->count() == 1) {
+            return redirect('usuarios')->withErrors([
+                'message' => 'Não é permitido excluir o usuário selecionado.',
+            ]);
+        }
+
         $usuario = User::find($id);
         $usuario->delete();
